Extract the signalling lock into its own support module
The tee-profile concurrency support module still exceeded the 400-line
file budget (425 lines) after the backend-availability split, bundling the
instrumented lock primitive with the worker/selector harness.

- Move the self-contained `_RLockLike` protocol and `_SignallingRLock`
  wrapper into `cuprum/unittests/_tee_profile_signalling_lock.py`; nothing
  in the harness referenced them, so no back-imports are needed.
- Drop them and the now-unused `types` import from
  `_tee_profile_concurrency_support.py`, leaving it at 345 lines focused on
  the worker plumbing, selectors, and race harness.
- Point `test_tee_profile_worker_concurrency.py` at the new lock module.
- Regenerate the wheel-layout snapshot and note the lock module in the
  developers' guide.
